move version info HTTP request into associated check
