+++ edgelet/iotedge/src/check/mod.rs
@@ -6,19 +6,15 @@ use std::path::PathBuf;
 
 use failure::Fail;
 use failure::{self, ResultExt};
-use futures::{future, Future, Stream};
 
 use edgelet_docker::Settings;
-use edgelet_http::client::ClientImpl;
-use edgelet_http::MaybeProxyClient;
 
 use aziotctl_common::{
     CheckOutputSerializable, CheckOutputSerializableStreaming, CheckResultSerializable,
     CheckResultsSerializable, CheckerMetaSerializable,
 };
 
-use crate::error::{Error, ErrorKind, FetchLatestVersionsReason};
-use crate::LatestVersions;
+use crate::error::{Error, ErrorKind};
 
 mod additional_info;
 use self::additional_info::AdditionalInfo;
@@ -40,7 +36,7 @@ pub struct Check {
     diagnostics_image_name: String,
     dont_run: BTreeSet<String>,
     aziot_edged: PathBuf,
-    latest_versions: Result<super::LatestVersions, Option<Error>>,
+    expected_aziot_edged_version: Option<String>,
     output_format: OutputFormat,
     verbose: bool,
     warnings_as_errors: bool,
@@ -98,121 +94,26 @@ impl Check {
         warnings_as_errors: bool,
         aziot_bin: std::ffi::OsString,
         iothub_hostname: Option<String>,
-    ) -> impl Future<Item = Self, Error = Error> + Send {
-        let latest_versions = if let Some(expected_aziot_edged_version) =
-            expected_aziot_edged_version
-        {
-            future::Either::A(future::ok::<_, Error>(LatestVersions {
-                aziot_edged: expected_aziot_edged_version,
-            }))
-        } else {
-            let proxy = std::env::var("HTTPS_PROXY")
-                .ok()
-                .or_else(|| std::env::var("https_proxy").ok())
-                .map(|proxy| proxy.parse::<hyper::Uri>())
-                .transpose()
-                .context(ErrorKind::FetchLatestVersions(
-                    FetchLatestVersionsReason::CreateClient,
-                ));
-            let hyper_client = proxy.and_then(|proxy| {
-                MaybeProxyClient::new(proxy, None, None).context(ErrorKind::FetchLatestVersions(
-                    FetchLatestVersionsReason::CreateClient,
-                ))
-            });
-            let hyper_client = match hyper_client {
-                Ok(hyper_client) => hyper_client,
-                Err(err) => {
-                    return future::Either::A(future::err(err.into()));
-                }
-            };
+    ) -> Check {
+        Check {
+            config_file,
+            container_engine_config_path,
+            diagnostics_image_name,
+            dont_run,
+            aziot_edged,
+            expected_aziot_edged_version,
+            output_format,
+            verbose,
+            warnings_as_errors,
+            aziot_bin,
 
-            let request = hyper::Request::get("https://aka.ms/latest-iotedge-stable")
-                .body(hyper::Body::default())
-                .expect("can't fail to create request");
+            additional_info: AdditionalInfo::new(),
 
-            future::Either::B(
-                hyper_client
-                    .call(request)
-                    .then(|response| -> Result<_, Error> {
-                        let response = response.context(ErrorKind::FetchLatestVersions(
-                            FetchLatestVersionsReason::GetResponse,
-                        ))?;
-                        Ok(response)
-                    })
-                    .and_then(move |response| match response.status() {
-                        hyper::StatusCode::MOVED_PERMANENTLY => {
-                            let uri = response
-                                .headers()
-                                .get(hyper::header::LOCATION)
-                                .ok_or(ErrorKind::FetchLatestVersions(
-                                    FetchLatestVersionsReason::InvalidOrMissingLocationHeader,
-                                ))?
-                                .to_str()
-                                .context(ErrorKind::FetchLatestVersions(
-                                    FetchLatestVersionsReason::InvalidOrMissingLocationHeader,
-                                ))?;
-                            let request = hyper::Request::get(uri)
-                                .body(hyper::Body::default())
-                                .expect("can't fail to create request");
-                            Ok(hyper_client.call(request).map_err(|err| {
-                                err.context(ErrorKind::FetchLatestVersions(
-                                    FetchLatestVersionsReason::GetResponse,
-                                ))
-                                .into()
-                            }))
-                        }
-                        status_code => Err(ErrorKind::FetchLatestVersions(
-                            FetchLatestVersionsReason::ResponseStatusCode(status_code),
-                        )
-                        .into()),
-                    })
-                    .flatten()
-                    .and_then(|response| -> Result<_, Error> {
-                        match response.status() {
-                            hyper::StatusCode::OK => {
-                                Ok(response.into_body().concat2().map_err(|err| {
-                                    err.context(ErrorKind::FetchLatestVersions(
-                                        FetchLatestVersionsReason::GetResponse,
-                                    ))
-                                    .into()
-                                }))
-                            }
-                            status_code => Err(ErrorKind::FetchLatestVersions(
-                                FetchLatestVersionsReason::ResponseStatusCode(status_code),
-                            )
-                            .into()),
-                        }
-                    })
-                    .flatten()
-                    .and_then(|body| {
-                        Ok(serde_json::from_slice(&body).context(
-                            ErrorKind::FetchLatestVersions(FetchLatestVersionsReason::GetResponse),
-                        )?)
-                    }),
-            )
-        };
-
-        future::Either::B(latest_versions.then(move |latest_versions| {
-            Ok(Check {
-                config_file,
-                container_engine_config_path,
-                diagnostics_image_name,
-                dont_run,
-                aziot_edged,
-                latest_versions: latest_versions.map_err(Some),
-                output_format,
-                verbose,
-                warnings_as_errors,
-                aziot_bin,
-
-                additional_info: AdditionalInfo::new(),
-
-                iothub_hostname,
-                settings: None,
-                docker_host_arg: None,
-                docker_server_version: None,
-            })
-        }))
+            iothub_hostname,
+            settings: None,
+            docker_host_arg: None,
+            docker_server_version: None,
+        }
     }
 
     pub fn print_list(aziot_bin: std::ffi::OsString) -> Result<(), Error> {
@@ -751,21 +652,19 @@ mod tests {
                 filename,
             );
 
-            let mut check = runtime
-                .block_on(Check::new(
-                    config_file.into(),
-                    "daemon.json".into(), // unused for this test
-                    "mcr.microsoft.com/azureiotedge-diagnostics:1.0.0".to_owned(), // unused for this test
-                    Default::default(),
-                    Some("1.0.0".to_owned()),  // unused for this test
-                    "aziot-edged".into(),      // unused for this test
-                    super::OutputFormat::Text, // unused for this test
-                    false,
-                    false,
-                    "".into(), // unused for this test
-                    None,
-                ))
-                .unwrap();
+            let mut check = Check::new(
+                config_file.into(),
+                "daemon.json".into(), // unused for this test
+                "mcr.microsoft.com/azureiotedge-diagnostics:1.0.0".to_owned(), // unused for this test
+                Default::default(),
+                Some("1.0.0".to_owned()),  // unused for this test
+                "aziot-edged".into(),      // unused for this test
+                super::OutputFormat::Text, // unused for this test
+                false,
+                false,
+                "".into(), // unused for this test
+                None,
+            );
 
             match WellFormedConfig::default().execute(&mut check, &mut runtime) {
                 CheckResult::Ok => (),
@@ -814,21 +713,19 @@ mod tests {
                 filename,
             );
 
-            let mut check = runtime
-                .block_on(Check::new(
-                    config_file.into(),
-                    "daemon.json".into(), // unused for this test
-                    "mcr.microsoft.com/azureiotedge-diagnostics:1.0.0".to_owned(), // unused for this test
-                    Default::default(),
-                    Some("1.0.0".to_owned()),  // unused for this test
-                    "aziot-edged".into(),      // unused for this test
-                    super::OutputFormat::Text, // unused for this test
-                    false,
-                    false,
-                    "".into(), // unused for this test
-                    None,
-                ))
-                .unwrap();
+            let mut check = Check::new(
+                config_file.into(),
+                "daemon.json".into(), // unused for this test
+                "mcr.microsoft.com/azureiotedge-diagnostics:1.0.0".to_owned(), // unused for this test
+                Default::default(),
+                Some("1.0.0".to_owned()),  // unused for this test
+                "aziot-edged".into(),      // unused for this test
+                super::OutputFormat::Text, // unused for this test
+                false,
+                false,
+                "".into(), // unused for this test
+                None,
+            );
 
             match WellFormedConfig::default().execute(&mut check, &mut runtime) {
                 CheckResult::Ok => (),
@@ -885,21 +782,19 @@ mod tests {
             filename,
         );
 
-        let mut check = runtime
-            .block_on(Check::new(
-                config_file.into(),
-                "daemon.json".into(), // unused for this test
-                "mcr.microsoft.com/azureiotedge-diagnostics:1.0.0".to_owned(), // unused for this test
-                Default::default(),
-                Some("1.0.0".to_owned()),  // unused for this test
-                "aziot-edged".into(),      // unused for this test
-                super::OutputFormat::Text, // unused for this test
-                false,
-                false,
-                "".into(), // unused for this test
-                None,
-            ))
-            .unwrap();
+        let mut check = Check::new(
+            config_file.into(),
+            "daemon.json".into(), // unused for this test
+            "mcr.microsoft.com/azureiotedge-diagnostics:1.0.0".to_owned(), // unused for this test
+            Default::default(),
+            Some("1.0.0".to_owned()),  // unused for this test
+            "aziot-edged".into(),      // unused for this test
+            super::OutputFormat::Text, // unused for this test
+            false,
+            false,
+            "".into(), // unused for this test
+            None,
+        );
 
         match WellFormedConfig::default().execute(&mut check, &mut runtime) {
             CheckResult::Failed(err) => {
@@ -932,21 +827,19 @@ mod tests {
             filename,
         );
 
-        let mut check = runtime
-            .block_on(super::Check::new(
-                config_file.into(),
-                "daemon.json".into(), // unused for this test
-                "mcr.microsoft.com/azureiotedge-diagnostics:1.0.0".to_owned(), // unused for this test
-                Default::default(),
-                Some("1.0.0".to_owned()),  // unused for this test
-                "aziot-edged".into(),      // unused for this test
-                super::OutputFormat::Text, // unused for this test
-                false,
-                false,
-                "".into(), // unused for this test
-                None,
-            ))
-            .unwrap();
+        let mut check = super::Check::new(
+            config_file.into(),
+            "daemon.json".into(), // unused for this test
+            "mcr.microsoft.com/azureiotedge-diagnostics:1.0.0".to_owned(), // unused for this test
+            Default::default(),
+            Some("1.0.0".to_owned()),  // unused for this test
+            "aziot-edged".into(),      // unused for this test
+            super::OutputFormat::Text, // unused for this test
+            false,
+            false,
+            "".into(), // unused for this test
+            None,
+        );
 
         match WellFormedConfig::default().execute(&mut check, &mut runtime) {
             CheckResult::Ok => (),
